🔧 ПРАВИЛЬНОЕ ИСПРАВЛЕНИЕ ЛАЙАУТА ШКОЛЫ ПО ПРИМЕРУ LOGIN
🎯 АНАЛИЗ ПРОБЛЕМ:
1. 🖼️ ФОТО НЕ ЗАГРУЖАЛОСЬ: В main ветке используется logoUrl а не imageUrl
2. 🏗️ ЛАЙАУТ НЕ РАБОТАЛ: Группы (school) не сбрасывали основной лайаут
3. 🔐 КНОПКА НЕ РАБОТАЛА: AdminModal был в неправильном месте

✅ РЕШЕНИЕ ПО ПРИМЕРУ LOGIN:
📁 СТРУКТУРА:
   src/routes/school/
   ├── +layout@.svelte     ← Сбрасывает ВСЕ лайауты (как login)
   ├── +page.svelte        ← Hero + превью модулей
   └── +page.ts            ← Загрузчик данных

🔥 ИСПРАВЛЕНИЯ:
1. 🖼️ ИЗОБРАЖЕНИЯ: Используем logoUrl (как в main)
   - Hero background: logoUrl
   - Hero logo: logoUrl
   - Debug логи: 'Hero background image loaded from logoUrl'
   - Placeholder если нет logoUrl

2. 🏗️ ЛАЙАУТ: +layout@.svelte сбрасывает все лайауты
   - Собственный навбар с брендом школы
   - AdminLoginModal интегрирован
   - Toast контейнер
   - Responsive дизайн

3. 🔐 АДМИН КНОПКА: Полная интеграция
   - openAdminModal() с debug логами
   - AdminLoginModal bind:open
   - Разные состояния кнопки

4. 📊 ЗАГРУЗКА ДАННЫХ:
   - +page.ts проверяет и загружает schoolData
   - Debug логи на каждом этапе
   - onMount дополнительная проверка

🧪 ТЕСТИРУЙТЕ:
- http://localhost:5173/school - лендинг БЕЗ сайдбара
- Кнопка 'Войти как админ' + console.log 'School Layout:'
- Фото школы загружается из logoUrl
