+++ src/routes/school/+page.ts
@@ -7,7 +7,15 @@ export async function load() {
 	// Загружаем данные школы если они не загружены
 	const currentAuth = get(authStore);
 	
+	console.log('🏫 School +page.ts load, currentAuth:', {
+		hasToken: !!currentAuth.token,
+		hasSchoolData: !!currentAuth.schoolData,
+		logoUrl: currentAuth.schoolData?.logoUrl,
+		schoolId: currentAuth.schoolId
+	});
+	
 	if (!currentAuth.schoolData && currentAuth.token) {
+		console.log('🏫 Loading school data from +page.ts...');
 		await authStore.loadSchoolData();
 	}
